requirements.txt
alembic==1.15.1
anyio==3.7.1
archspec==0.2.4
argcomplete==3.6.2
asgiref==3.11.0
asttokens==3.0.1
attrs==24.2.0
babel==2.17.0
Beaker==1.12.1
beautifulsoup4==4.14.2
blinker==1.9.0
blivet==3.11.0
blivet-gui==2.6.0
boltons==24.0.0
Brlapi==0.8.5
Brotli==1.1.0
certifi==2023.5.7
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.3
conda==24.9.1
conda-libmamba-solver==23.11.1
conda-package-handling==2.3.0
conda_package_streaming==0.10.0
contourpy==1.3.3
coverage==7.6.12
crypt_r==3.13.1
cryptography==43.0.0
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
debugpy==1.8.17
decorator==5.2.1
dill==0.4.1
distro==1.9.0
Django==5.2.8
djangorestframework==3.16.1
dnf==4.24.0
executing==2.2.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.20.3
Flask==3.1.0
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WhooshAlchemy==0.56
Flask-WTF==1.2.2
fonttools==4.60.1
fros==1.1
frozendict==2.4.4
fsspec==2026.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
image==1.5.33
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.11.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.22
libdnf==0.75.0
libmambapy==1.5.8
llvmlite==0.46.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
menuinst==2.0.0
mpmath==1.3.0
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.6.1
nftables==0.1
nlopt==2.10.0
numba==0.63.1
numpy==2.2.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
olefile==0.47
packaging==24.1
pandas==2.2.3
parso==0.8.5
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
productmd==1.45
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.25.1
pycosat==0.6.6
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.23.0
pycups==2.0.4
pydot==4.0.1
pyenchant==3.2.2
Pygments==2.19.2
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.58
pylatexenc==2.10
pyOpenSSL==24.2.1
pyparsing==3.2.5
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
pytz==2025.1
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
pyynl @ file:///builddir/build/BUILD/kernel-6.17.10-build/kernel-6.17.10/linux-6.17.10-100.fc41.x86_64/tools/net/ynl
pyzmq==27.1.0
qiskit==2.3.0
qiskit-aer==0.17.2
qiskit-machine-learning==0.9.0
rapidfuzz==3.5.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.10
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.19.1
rpm==4.20.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.12
rustworkx==0.17.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
sparse==0.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.6.0
sympy==1.14.0
systemd-python==235
Tempita==0.5.2
threadpoolctl==3.5.0
torch==2.10.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.14
websockets==12.0
Werkzeug==3.1.3
Whoosh==2.7.4
WTForms==3.2.1
xkbregistry==0.3
yt-dlp==2025.6.30
zstandard==0.23.0
